+++ Bezhkov/Bulk-Thumbnail-Creator/ConsoleApp14/ConsoleApp14.csproj
@@ -49,6 +49,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Choise.cs" />
+    <Compile Include="Images.cs" />
     <Compile Include="ManagerProgramm.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
